Renamed "clients"-folder to "client". No other modifications
