+++ src/components/ErrorPrompt.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { Notification } from 'hds-react';
 
-import { useClientErrorDetection } from '../clients/client';
-import { ClientErrorObject, ClientError } from '../clients';
+import { useClientErrorDetection } from '../client/client';
+import { ClientErrorObject, ClientError } from '../client';
 import styles from './styles.module.css';
 
 const ErrorPrompt = (
