+++ src/components/Profile.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { Button } from 'hds-react';
-import { ApiAccessTokenActions } from '../clients/client';
+import { ApiAccessTokenActions } from '../client/client';
 import { ProfileDataType, useProfile } from '../profile/profile';
 import { ApiAccessTokenContext } from './ApiAccessTokenProvider';
 import styles from './styles.module.css';
