+++ src/profile/profile.ts
@@ -5,8 +5,8 @@ import to from 'await-handler';
 import { ApolloError } from '@apollo/client';
 import { loader } from 'graphql.macro';
 import { useCallback, useContext, useEffect, useState } from 'react';
-import { FetchStatus } from '../clients/client';
-import { getClient } from '../clients/oidc-react';
+import { FetchStatus } from '../client/client';
+import { getClient } from '../client/oidc-react';
 import {
   GraphQLClient,
   createGraphQLClient,
